Warn that OneDrive for Business URL's will change
The OneDrive for Business (ODB) URL is derived from a user's UPN. When doing a "Set as Primary", this changes the user's UPN in M365 which hence also causes the ODB URL to be changed. The downstream affect is that there can be issues for other applications that depend on the URL. For example, if a user has a OneNote file located in ODB, then this file may stop syncing.

This proposed update is to provide a warning that the OneDrive URL will change when performing a UPN rename.